--- a/src/kernels/MIOpenTensorKernelsHip.cpp
+++ b/src/kernels/MIOpenTensorKernelsHip.cpp
@@ -232,114 +232,80 @@ extern "C" __global__ void Op3dTensorGeneric(const MIOPEN_TYPE* a,
 
 #ifdef USE_4D_TENSOR_GENERIC
 // NCHW
-extern "C" __global__ void Op4dTensorGeneric(MIOPEN_TYPE* a,
-                                             const int a_nstride,
-                                             const int a_cstride,
-                                             const int a_hstride,
-                                             MIOPEN_TYPE* b,
-                                             const int b_c,
-                                             const int b_h,
-                                             const int b_w,
-                                             const int b_nstride,
-                                             const int b_cstride,
-                                             const int b_hstride,
+extern "C" __global__ void Op4dTensorGeneric(const MIOPEN_TYPE* a,
+                                             const MIOPEN_TYPE* b,
                                              MIOPEN_TYPE* c,
-                                             const int c_c,
-                                             const int c_h,
-                                             const int c_w,
-                                             const int c_nstride,
-                                             const int c_cstride,
-                                             const int c_hstride,
+                                             const uint64_t Aoffset,
+                                             const uint64_t Boffset,
+                                             const uint64_t Coffset,
+                                             const uint32_t b_c,
+                                             const uint32_t b_h,
+                                             const uint32_t b_w,
+                                             const uint32_t c_c,
+                                             const uint32_t c_h,
+                                             const uint32_t c_w,
+                                             const uint32_t a_nstride,
+                                             const uint32_t a_cstride,
+                                             const uint32_t a_hstride,
+                                             const uint32_t a_wstride,
+                                             const uint32_t b_nstride,
+                                             const uint32_t b_cstride,
+                                             const uint32_t b_hstride,
+                                             const uint32_t b_wstride,
+                                             const uint32_t c_nstride,
+                                             const uint32_t c_cstride,
+                                             const uint32_t c_hstride,
+                                             const uint32_t c_wstride,
                                              const MIOPEN_TYPE alpha0,
                                              const MIOPEN_TYPE alpha1,
                                              const MIOPEN_TYPE beta,
-                                             const unsigned int bitmap,
-                                             const int work_per_wg,
-                                             const long Aoffset,
-                                             const long Boffset,
-                                             const long Coffset,
-                                             const int num_wg)
+                                             const uint32_t total_work,
+                                             const bool use_beta)
 {
-    int gid = blockIdx.x;
+    const MIOPEN_TYPE* a_off = a + Aoffset;
+    const MIOPEN_TYPE* b_off = b + Boffset;
+    MIOPEN_TYPE* c_off       = c + Coffset;
 
-    MIOPEN_TYPE* a_off = a + Aoffset;
-    MIOPEN_TYPE* b_off = b + Boffset;
-    MIOPEN_TYPE* c_off = c + Coffset;
+    auto gid          = blockIdx.x * blockDim.x + threadIdx.x;
+    const auto* a_ptr = a_off + (gid / (c_c * c_h * c_w)) * a_nstride +
+                        ((gid % (c_c * c_h * c_w)) / (c_h * c_w)) * a_cstride +
+                        ((gid % (c_h * c_w)) / c_w) * a_hstride + (gid % c_w) * a_wstride;
+    auto* c_ptr = c_off + (gid / (c_c * c_h * c_w)) * c_nstride +
+                  ((gid % (c_c * c_h * c_w)) / (c_h * c_w)) * c_cstride +
+                  ((gid % (c_h * c_w)) / c_w) * c_hstride + (gid % c_w) * c_wstride;
 
-// MIOPEN_TYPE operand = b[gid + Boffset];
-// num_wg: the number of workgroups should be launched
-// MAX_NUM_WG: the maximum number of workgroups actually launched
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wfloat-equal"
-    if(beta == static_cast<MIOPEN_TYPE>(0))
-#pragma clang diagnostic pop
+    const auto step   = gridDim.x * blockDim.x;
+    const auto a_step = (step / (c_c * c_h * c_w)) * a_nstride +
+                        ((step % (c_c * c_h * c_w)) / (c_h * c_w)) * a_cstride +
+                        ((step % (c_h * c_w)) / c_w) * a_hstride + (step % c_w) * a_wstride;
+    const auto c_step = (step / (c_c * c_h * c_w)) * c_nstride +
+                        ((step % (c_c * c_h * c_w)) / (c_h * c_w)) * c_cstride +
+                        ((step % (c_h * c_w)) / c_w) * c_hstride + (step % c_w) * c_wstride;
+
+    const auto c_end = c_off + total_work * c_nstride;
+    while(c_ptr < c_end)
     {
-        for(; gid < num_wg; gid += MAX_NUM_WG)
-        {
-            int lid = threadIdx.x;
+        const auto* b_ptr = b_off;
+        if(b_nstride != 0)
+            b_ptr += (gid / (b_c * b_h * b_w)) * b_nstride;
 
-            int o_h_div = (bitmap & (1 << 0)) ? 1 : c_w;
-            int o_c_div = o_h_div * ((bitmap & (1 << 1)) ? 1 : c_h);
-            int o_n_div = o_c_div * ((bitmap & (1 << 2)) ? 1 : c_c);
+        if(b_cstride != 0)
+            b_ptr += ((gid % (b_c * b_h * b_w)) / (b_h * b_w)) * b_cstride;
 
-            int o_w_gid_off = gid % b_w;
-            int o_h_gid_off = (gid / b_w) % b_h;
-            int o_c_gid_off = (gid / b_w / b_h) % b_c;
-            int o_n_gid_off = (gid / b_w / b_h) / b_c;
+        if(b_hstride != 0)
+            b_ptr += ((gid % (b_h * b_w)) / b_w) * b_hstride;
 
-            int bindex = o_n_gid_off * b_nstride + o_c_gid_off * b_cstride +
-                         o_h_gid_off * b_hstride + o_w_gid_off;
-            MIOPEN_TYPE operand = b_off[bindex] * alpha1;
+        if(b_wstride != 0)
+            b_ptr += (gid % b_w) * b_wstride;
 
-            while(lid < work_per_wg)
-            {
-                int o_w = (bitmap & (1 << 0)) ? o_w_gid_off : lid % c_w;
-                int o_h = (bitmap & (1 << 1)) ? o_h_gid_off : (lid / o_h_div) % c_h;
-                int o_c = (bitmap & (1 << 2)) ? o_c_gid_off : (lid / o_c_div) % c_c;
-                int o_n = (bitmap & (1 << 3)) ? o_n_gid_off : lid / o_n_div;
+        auto a_val = *a_ptr;
+        auto b_val = *b_ptr;
+        auto c_val = use_beta ? *c_ptr : static_cast<MIOPEN_TYPE>(0);
+        *c_ptr     = MIOPEN_TENSOR_OP(a_val * alpha0, b_val * alpha1) + c_val * beta;
 
-                int aindex    = o_n * a_nstride + o_c * a_cstride + o_h * a_hstride + o_w;
-                int cindex    = o_n * c_nstride + o_c * c_cstride + o_h * c_hstride + o_w;
-                c_off[cindex] = MIOPEN_TENSOR_OP(a_off[aindex] * alpha0, operand);
-
-                lid += blockDim.x;
-            }
-        }
-    }
-    else
-    {
-        for(; gid < num_wg; gid += MAX_NUM_WG)
-        {
-            int lid = threadIdx.x;
-
-            int o_h_div = (bitmap & (1 << 0)) ? 1 : c_w;
-            int o_c_div = o_h_div * ((bitmap & (1 << 1)) ? 1 : c_h);
-            int o_n_div = o_c_div * ((bitmap & (1 << 2)) ? 1 : c_c);
-
-            int o_w_gid_off = gid % b_w;
-            int o_h_gid_off = (gid / b_w) % b_h;
-            int o_c_gid_off = (gid / b_w / b_h) % b_c;
-            int o_n_gid_off = (gid / b_w / b_h) / b_c;
-
-            int bindex = o_n_gid_off * b_nstride + o_c_gid_off * b_cstride +
-                         o_h_gid_off * b_hstride + o_w_gid_off;
-            MIOPEN_TYPE operand = b_off[bindex] * alpha1;
-
-            while(lid < work_per_wg)
-            {
-                int o_w = (bitmap & (1 << 0)) ? o_w_gid_off : lid % c_w;
-                int o_h = (bitmap & (1 << 1)) ? o_h_gid_off : (lid / o_h_div) % c_h;
-                int o_c = (bitmap & (1 << 2)) ? o_c_gid_off : (lid / o_c_div) % c_c;
-                int o_n = (bitmap & (1 << 3)) ? o_n_gid_off : lid / o_n_div;
-
-                int aindex = o_n * a_nstride + o_c * a_cstride + o_h * a_hstride + o_w;
-                int cindex = o_n * c_nstride + o_c * c_cstride + o_h * c_hstride + o_w;
-                c_off[cindex] =
-                    MIOPEN_TENSOR_OP(a_off[aindex] * alpha0, operand) + beta * c_off[cindex];
-
-                lid += blockDim.x;
-            }
-        }
+        a_ptr += a_step;
+        c_ptr += c_step;
+        gid += step;
     }
 }