… src/kernels/MIOpenTensorKernels.cl
… test/gtest/tensor_4d_generic_ocl_hip.cpp