## Proposed changes
Reopening PR https://github.com/ROCm/MIOpen/pull/3458

Tested again performance of these kernels but with other measurements, instead of time comparison I calculated useful calculations per second (GFLOPs) and bytes transferred to/from memory (GBs)

Tested with tests generated in `tensor_4d_generic_ocl_hip.cpp` file, only packed tensors, float data type,, size from 32MB to 4GB per tensor, tested on gfx1030 (Radeon RX6800XT), comparison performed on applicable B tensor dimensions

Comparison with old Op4dTensorGeneric

|    | Old kernel | New kernel | Speed-up |
| - | ----------- | ------------ | ------------ |
| GFLOPs | 3.187 | 193.437 | x60.7 |
| GBs | 11.135 | 494.846 |  x44.44 |

Comparison with Op4dTensorLite

|    | Op4dTensorLIte | New Op4dTensorGeneric | Speed-up |
| - | ----------- | ------------ | ------------ |
| GFLOPs | 138.89 | 116.723 | x0.84 |
| GBs | 104.65 | 368.346 |  x3.52 |

Comparison with OpTensorFwdBias

|    | OpTensorFwdBias | New Op4dTensorGeneric | Speed-up |
| - | ----------- | ------------ | ------------ |
| GFLOPs | 60.1 | 204.468 | x3.4 |
| GBs | 192.591 | 500.374 |  x2.6 |

Comparison with OpTensorLeadingOnes

|    | OpTensorLeadingOnes | New Op4dTensorGeneric| Speed-up |
| - | ----------- | ------------ | ------------ |
| GFLOPs | 116.485 | 209.944 | x1.8 |
| GBs | 382.389 | 498.089 |  x1.3 |


## Checklist

Please put an `x` into the boxes that apply. You can also fill these out after creating the PR. If you're not sure, please don't hesitate to ask.

- [ x ] I have added automated tests relevant to the introduced functionality
- [ x ] I have sufficient test coverage for the changes, and code coverage hasn't decreased as a result of my PR
- [ x ] I have ran the tests, and they are all passing locally
- [ x ] I have added relevant documentation for the changes
- [ x ] I have removed the stale documentation which is no longer relevant after this pull request
- [ x ] I have ran `make format` & `make check_format` to ensure the changes have been formatted